optimiser signal state using exit values
The optimisers signal state using exit values. An exit value of 1
indicates a new paramter set is required and an exit value of 2
indicates that the optimiser is waiting for actual function results.

The simple example scripts are combined into a single script. Use
command line option to select between nlopt and dfols optimiser.